+++ lib/protocol/command_table.h
@@ -9,6 +9,7 @@
 #include <memory>
 #include <unordered_map>
 #include <cassert>
+
 constexpr uint32_t PROTOCOL_VERS = 2;
 
 using identifier_t = int32_t;
