+++ lib/protocol/protocol.h
@@ -1,0 +1,19 @@
+#ifndef PROTOCOL_H
+#define PROTOCOL_H
+
+#include <iostream>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <cstdint>
+#include <cstdio>
+
+#include <request/input_request.h>
+#include <response/input_response.h>
+#include <request/transport_request.h>
+#include <response/transport_response.h>
+#include <request/control_request.h>
+
+
+
+#endif // PROTOCOL_H
