--- src/protocol/protocol.cpp
@@ -1,2 +1,0 @@
-#include <protocol/protocol.h>
-
