--- src/protocol/protocol.h
@@ -1,19 +1,0 @@
-#ifndef PROTOCOL_H
-#define PROTOCOL_H
-
-#include <iostream>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring>
-#include <cstdint>
-#include <cstdio>
-
-#include <protocol/request/input_request.h>
-#include <protocol/response/input_response.h>
-#include <protocol/request/transport_request.h>
-#include <protocol/response/transport_response.h>
-#include <protocol/request/control_request.h>
-
-
-
-#endif // PROTOCOL_H
