--- src/protocol/request/input_request.cpp
@@ -1,17 +1,0 @@
-#include <protocol/request/input_request.h>
-
-std::ostream& operator<<(std::ostream& os, const RegistrationRequest& request) {
-    os << "protocol version=" << request.get_protocol_version()
-        << ", type command=" << static_cast<uint16_t>(request.get_type_data())
-        << ", login=" << request.get_login()
-        << ", password=" << request.get_password();
-    return os;
-}
-
-std::ostream& operator<<(std::ostream& os, const AutorisationRequest& request) {
-    os << "protocol version=" << request.get_protocol_version()
-        << ", type command=" << static_cast<uint16_t>(request.get_type_data())
-        << ", login=" << request.get_login()
-        << ", password=" << request.get_password();
-    return os;
-}
