--- src/protocol/request/transport_request.cpp
@@ -1,1 +1,0 @@
-#include <protocol/request/transport_request.h>
