+++ src/server/channel/iroom.h
@@ -5,7 +5,7 @@
 #include <string_view>
 
 #include <server/connection/isubscriber.h>
-#include <protocol/protocol.h>
+#include <protocol.h>
