+++ src/server/connection/connection.h
@@ -7,7 +7,7 @@
 
 #include <server/connection/isubscriber.h>
 #include <server/channel/channels_manager.h>
-#include <protocol/protocol.h>
+#include <protocol.h>
 
 class Connection : public ISubscriber, public std::enable_shared_from_this<Connection>
 {
