+++ src/server/connection/isubscriber.h
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <memory>
 #include <string>
-#include <protocol/protocol.h>
+#include <protocol.h>
