+++ src/storage/database.h
@@ -1,7 +1,7 @@
 #ifndef DATABASE_H
 #define DATABASE_H
 
-#include <protocol/protocol.h>
+#include <protocol.h>
 #include <deque>
 #include <sqlite3.h>
